move config related function
